init commit

+++ api/app/models_proxy.py
@@ -120,7 +120,7 @@ class Ad(Base):
     last_seen_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
     last_refresh_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
 
-    reach: Mapped[int | None] = mapped_column(Integer, nullable=True, index=True)
+    reach: Mapped[int | None] = mapped_column(BigInteger, nullable=True, index=True)
     reach_breakdown: Mapped[dict | None] = mapped_column(JSONB, nullable=True)
 
     raw_data: Mapped[dict | None] = mapped_column(JSON, nullable=True)
